Search component moved to common
