+++ src/common/Search/styled.js
